module1-validator/requirements.txt
azure-functions
redis
azure-eventhub
azure-storage-blob
python-dotenv